+++ starter-code/routes/index.js
@@ -1,9 +1,27 @@
 const express = require('express');
 const router  = express.Router();
 
+const passportRouter = express.Router();
+const passport = require("passport");
+const User = require("../models/user");
+const bcrypt = require("bcrypt");
+const bcryptSalt = 10;
+const ensureLogin = require("connect-ensure-login");
+
+
 /* GET home page */
 router.get('/', (req, res, next) => {
   res.render('index');
 });
+
+router.get('/login', (req, res, next) => {
+  res.render('passport/login.hbs');
+});
+router.post('/login', passport.authenticate("local", {
+  successRedirect: "/",
+  failureRedirect: "/login",
+  failureFlash: true,
+  passReqToCallback: true
+}));
 
 module.exports = router;
