+++ starter-code/views/passport/private.hbs
